KOGTIO-5804: prefix package as org.kie.kogito - dmn-[quarkus|springboot]-example
